Fixed bug in palette loading, made cycling and lights work together.
Routine to load dac was supposed to start from the second page, because fades
happen on the first page. It was loading _to_ there, but it was loading _from_
colors meant for the first row.

The last 8 color pages were unused, now they handle cycling in lights-off mode.

An additonal tileset was added in order to better demonstrate the available
features.